third_party/blink/web_tests/html/details_summary/details-nested-1-expected.txt

CONSOLE ERROR: An interactive element was found within a <summary> element. These elements will not consistently be accessible to people navigating by keyboard or using assistive technology.
